pull_step using make_files_same() was a totally stupid thing to do. Move back to using patches.
I suspect that the rebase version of finish() is the right thing, but
until I know that, sticking with merge.

Ugh. This is horrific.

+++ welded/ops.py
@@ -51,8 +51,9 @@ def delete_seams(spec, base_obj, seams, base_commit):
 
     for s in seams:
         to_zap = os.path.join(spec.base_dir, s.dest)
-        print("W: Remove %s\n"%to_zap)
-        shutil.rmtree(to_zap)
+        if os.path.exists(to_zap):
+            print("W: Remove %s\n"%to_zap)
+            shutil.rmtree(to_zap)
         git.add_in_subdir(spec.base_dir, s.dest )
     
     # Now create the header for all this ..
@@ -126,6 +127,29 @@ def rewrite_diff(infile, cid, changes):
                     
 
     return (are_any, outfile)
+
+def list_files_involved(base_changes):
+    """
+    Returns a list of (letter, path) pairs.
+    """
+    r = re.compile(r'^:([0-9]+)\s+([0-9]+)\s+([0-9a-f]+)...\s+([0-9a-f]+)...\s+([A-Z])\s+(.*)$')
+    rv = [ ]
+    for b in base_changes:
+        m = r.match(b)
+        if (m is not None):
+            rv.append((m.group(5), m.group(6).strip()))
+    return rv
+    
+
+def list_files_by_attribute(base_changes, attr):
+    r = re.compile(r'^:([0-9]+)\s+([0-9]+)\s+([0-9a-f]+)...\s+([0-9a-f]+)...\s+([A-Z])\s+(.*)$')
+    rv = [ ]
+    for b in base_changes:
+        m = r.match(b)
+        if (m is not None) and (attr == m.group(5)):
+            rv.append(m.group(6).strip())
+    return rv
+
 
         
 def modify_seams(spec, base_obj, changes, old_commit, new_commit):
